Implement ubsan and fix undefined behavior

+++ kernel/include/bitset.h
@@ -3,11 +3,11 @@
 #include <stdint.h>
 
 static inline void bitset_set(uint64_t* bitset, size_t i) {
-  bitset[i / 64] |= 1 << (i % 64);
+  bitset[i / 64] |= (uint64_t) 1 << (i % 64);
 }
 static inline void bitset_clear(uint64_t* bitset, size_t i) {
-  bitset[i / 64] &= ~(1 << (i % 64));
+  bitset[i / 64] &= ~((uint64_t) 1 << (i % 64));
 }
 static inline size_t bitset_test(uint64_t* bitset, size_t i) {
-  return bitset[i / 64] & 1 << (i % 64);
+  return bitset[i / 64] & (uint64_t) 1 << (i % 64);
 }
